Hassan

+++ src/Controller/OutingController.php
@@ -8,6 +8,7 @@ use App\Repository\OutingRepository;
 use App\Service\OutingService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -65,10 +66,14 @@ final class OutingController extends AbstractController
     }
 
     #[Route('/{id}/edit', name: 'app_outing_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, OutingService $outingService, Outing $outing, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, Outing $outing, OutingService $outingService , EntityManagerInterface $entityManager, Security $security): Response
     {
         $form = $this->createForm(OutingType::class, $outing);
         $form->handleRequest($request);
+
+        if ($outing->getIdOrganizer()->getId() != $security->getUser()->getId()) {
+            return $this->redirectToRoute('main_home');
+        }
 
         if ($form->isSubmitted() && $form->isValid()) {
             $newOutingState = $outingService->calculateOutingState($outing);
@@ -86,13 +91,37 @@ final class OutingController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_outing_delete', methods: ['POST'])]
-    public function delete(Request $request, Outing $outing, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, Outing $outing, EntityManagerInterface $entityManager, Security $security): Response
     {
+        if ($outing->getIdOrganizer()->getId() != $security->getUser()->getId()) {
+            return $this->redirectToRoute('main_home');
+        }
+
         if ($this->isCsrfTokenValid('delete' . $outing->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($outing);
             $entityManager->flush();
         }
 
         return $this->redirectToRoute('app_outing_index', [], Response::HTTP_SEE_OTHER);
+    }
+
+    #[Route('/withdrew/{id}', name: 'app_outing_withdrew')]
+    public function withdrew (Outing $outing, Security $security, EntityManagerInterface $entityManager): Response
+    {
+        if(!$outing->getIdMember()->contains($security->getUser())) {
+            return $this->redirectToRoute('main_home');
+        }
+
+        if(strcasecmp($outing->getState(), "ouvert") == 0){
+            $outing->getIdMember()->removeElement($security->getUser());
+
+            if($outing->getRegistrationDeadline() < new \DateTime()){
+                $outing->setSlots($outing->getSlots() + 1);
+            }
+
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('main_home');
     }
 }
